backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.17.0
httpx==0.26.0
stripe==7.10.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

email-validator>=2.0.0
cryptography==41.0.7
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
slowapi==0.1.9
